Keep package build orchestration in one script
